requirements.txt
kafka
mysql-connector-python
pymysql
pandas
python-dotenv